Utils deleted. Functions passed to handlers.
