+++ config_data/config.py
@@ -7,8 +7,5 @@ else:
     load_dotenv()
 
 BOT_TOKEN = os.getenv("BOT_TOKEN")
-RAPID_API_KEY = os.getenv("RAPID_API_KEY")
-DEFAULT_COMMANDS = (
-    ("start", "Запустить бота"),
-    ("help", "Помощь по командам бота")
-)
+RAPID_API_KEY = os.getenv("API_KEY")
+
