+++ env.template
@@ -1,1 +1,2 @@
 'BOT_TOKEN' = 'your_token_here'
+'API_KEY' = 'your api key for rapidapi'
