+++ handlers/custom_handlers/lowprice.py
@@ -1,8 +1,27 @@
 from telebot.types import Message
+import json
 
 from loader import bot
-from utils.lowprice import lowprice
-from states.user_request import DBRequest, UserRequest
+import requests
+from states.user_request import UserRequest
+
+url = 'https://hotels4.p.rapidapi.com/properties/v2/list'
+sort_string = 'PRICE_LOW_TO_HIGH'
+
+
+def lowprice(headers, querystring):
+    """
+    Функция для нахождения топ самых дешевых отелей.
+    :return: list список самых дешевых отелей
+
+    """
+    response = requests.request('POST',
+                                url=url,
+                                headers=headers,
+                                params=querystring)
+
+    data = json.loads(response.text)
+    return 'Узнать топ самых дешевых отелей в городе'
 
 
 @bot.message_handler(commands=['lowprice'])
@@ -44,4 +63,3 @@ def print_result(message: Message):
                 print(data['quantity'])
             else:
                 print('Нет данных')
-
