--- utils/history.py
@@ -1,7 +1,0 @@
-def history():
-	"""
-	Функция для запроса истории поиска.
-	:return: list список предыдущих запросов
-
-	"""
-	return 'Узнать историю поиска отелей'
